✏️ [fix] MemberService 인터페이스와 Impl로 분리
